test(units): split project_units tests into tests.rs (module-size ratchet)
The inline test module pushed project_units/mod.rs to 419 lines, over the 400-line
module-size ratchet (rust/processing/tests/module_size_ratchet.rs). Move the tests
to project_units/tests.rs (exempt as an embedded test file, per the ratchet rule)
so the production module stays at 265 lines. No test or behavior change.
